refactor(ipfs): move cross-resource deletion workflows out of shared client
