+++ backend/src/modules/documents/application/delete-document.usecaste.ts
@@ -1,12 +1,14 @@
 import { Response } from 'express';
 
 import { checkPermissionForWorkspace } from '../../../shared/utility/permissions';
+import { deleteDocumentsAndAssociatedItems } from './delete-documents-and-associated-items';
 import { documentsIpfsRepository } from '../infrastructure/documents-ipfs.repository';
 
 export async function deleteDocument(documentId: string, email: string, res: Response) {
   const doc = await documentsIpfsRepository.getDocumentDetailsById(documentId);
   await checkPermissionForWorkspace(email, res, doc.meta.workspaceOrigin);
 
-  const result = await documentsIpfsRepository.deleteDocument(documentId);
+  const items = await documentsIpfsRepository.getEverythingByDocId(documentId);
+  const result = await deleteDocumentsAndAssociatedItems(items);
   return { result };
 }
